+++ src/datasources/results/query-handlers/query-results/QueryResultsDataSource.ts
@@ -3,14 +3,13 @@ import { ResultsDataSourceBase } from "datasources/results/ResultsDataSourceBase
 import { DataQueryRequest, DataFrameDTO, FieldType, LegacyMetricFindQueryOptions, MetricFindValue } from "@grafana/data";
 import { OutputType } from "datasources/results/types/types";
 import { defaultResultsQuery } from "datasources/results/defaultQueries";
+import { ExpressionTransformFunction, transformComputedFieldsQuery } from "core/query-builder.utils";
+import { ResultsQueryBuilderFieldNames } from "datasources/results/constants/ResultsQueryBuilder.constants";
 
 export class QueryResultsDataSource extends ResultsDataSourceBase {
   queryResultsUrl = this.baseUrl + '/v2/query-results';
-  queryResultsValuesUrl = this.baseUrl + '/v2/query-result-values';
 
   defaultQuery = defaultResultsQuery;
-
-  readonly partNumbersCache: string[] = [];
 
   async queryResults(
     filter?: string,
@@ -38,8 +37,17 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
     await this.getPartNumbers();
     await this.loadWorkspaces();
 
+    if (query.queryBy) {
+      query.queryBy = transformComputedFieldsQuery(
+        this.templateSrv.replace(query.queryBy, options.scopedVars),
+        this.resultsComputedDataFields,
+      );
+    }
+
+    const useTimeRangeFilter = this.getTimeRangeFilter(options, query.useTimeRange, query.useTimeRangeFor);
+
     const responseData = await this.queryResults(
-      this.getTimeRangeFilter(options, query.useTimeRange, query.useTimeRangeFor),
+      this.buildQueryFilter(query.queryBy, useTimeRangeFilter),
       query.orderBy,
       query.properties,
       query.recordCount,
@@ -84,20 +92,19 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
     }
   }
 
-  async getPartNumbers(): Promise<void> {
-    if (this.partNumbersCache.length > 0) {
-      return;
-    }
-    
-    const partNumbers = await this.post<string[]>(this.queryResultsValuesUrl, {
-      field: ResultsPropertiesOptions.PART_NUMBER,
-    }).catch(error => {
-      throw new Error(error);
-    });
-
-    partNumbers?.forEach(partNumber => this.partNumbersCache.push(partNumber));
-  }
-
+  /**
+   * A map linking each field name to its corresponding query transformation function.
+   * It dynamically processes and formats query expressions based on the field type.
+   */
+  readonly resultsComputedDataFields = new Map<string, ExpressionTransformFunction>(
+    Object.values(ResultsQueryBuilderFieldNames).map(field => [
+      field,
+      field === (ResultsQueryBuilderFieldNames.UPDATED_AT) || field === (ResultsQueryBuilderFieldNames.STARTED_AT)
+        ? this.timeFieldsQuery(field)
+        : this.multipleValuesQuery(field),
+    ])
+  );
+  
   async metricFindQuery(_query: ResultsVariableQuery, _options?: LegacyMetricFindQueryOptions): Promise<MetricFindValue[]> {
     return [];
   }